(PC-2555): BSR -> re-organized signin, signup folders
